refactor test directories
