+++ src/test/java/control/TempFilesFabricTest.java
@@ -1,6 +1,5 @@
-package temp_files_fabric;
+package control;
 
-import control.TempFilesFabric;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
